Descarga y lee los datos de archivos de estaciones meteorológicas

Descarga un archivo formato CSV con datos de una estación meteorológica específica y lo guarda en una ruta local. Si el archivo ya existe en la ruta especificada, la función lo lee directamente sin volver a descargarlo.

Usage

leer_estaciones(id_estacion, ruta)

Arguments

id_estacion

ID de la estación cuyos datos se quieren leer.

ruta

Ruta o directorio en donde se descargarán y leerán los datos

Value

Los datos en formato CSV para poder almacenarlos en una variable y utilizarlos posteriormente.

Details

Requiere los paquetes glue, readr y cli.

Examples

leer_estaciones("NH0437", ruta_temp1)
#> El archivo no existe en la ruta indicada. Descargando...
#> Descarga completada.
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)
#> Rows: 18929 Columns: 35
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (2): id, direccion_viento_1000cm
#> dbl  (22): temperatura_abrigo_150cm, temperatura_abrigo_150cm_maxima, temper...
#> lgl  (10): temperatura_suelo_10cm_media, temperatura_inte_5cm, temperatura_i...
#> date  (1): fecha
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Lectura completada. El dataset tiene 18929 filas y 35 columnas.
#> # A tibble: 18,929 × 35
#>    id     fecha      temperatura_abrigo_150cm temperatura_abrigo_150cm_maxima
#>    <chr>  <date>                        <dbl>                           <dbl>
#>  1 NH0437 1970-08-01                    23.3                             31.2
#>  2 NH0437 1970-08-02                    26.6                             32.1
#>  3 NH0437 1970-08-03                    17.4                             22.9
#>  4 NH0437 1970-08-04                    11.0                             11.9
#>  5 NH0437 1970-08-05                    11.1                             12.6
#>  6 NH0437 1970-08-06                    13.6                             15.1
#>  7 NH0437 1970-08-07                    10.8                             14.9
#>  8 NH0437 1970-08-08                     7.75                            13.7
#>  9 NH0437 1970-08-09                     8.2                             15.1
#> 10 NH0437 1970-08-10                    10.2                             18.4
#> # ℹ 18,919 more rows
#> # ℹ 31 more variables: temperatura_abrigo_150cm_minima <dbl>,
#> #   temperatura_intemperie_5cm_minima <dbl>,
#> #   temperatura_intemperie_50cm_minima <dbl>,
#> #   temperatura_suelo_5cm_media <dbl>, temperatura_suelo_10cm_media <lgl>,
#> #   temperatura_inte_5cm <lgl>, temperatura_intemperie_150cm_minima <lgl>,
#> #   humedad_suelo <lgl>, precipitacion_pluviometrica <dbl>, granizo <dbl>, …
leer_estaciones("NH0910", ruta_temp2)
#> El archivo no existe en la ruta indicada. Descargando...
#> Descarga completada.
#> Rows: 15553 Columns: 35
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (1): id
#> dbl   (5): temperatura_abrigo_150cm, temperatura_abrigo_150cm_maxima, temper...
#> lgl  (28): temperatura_intemperie_5cm_minima, temperatura_intemperie_50cm_mi...
#> date  (1): fecha
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Lectura completada. El dataset tiene 15553 filas y 35 columnas.
#> # A tibble: 15,553 × 35
#>    id     fecha      temperatura_abrigo_150cm temperatura_abrigo_150cm_maxima
#>    <chr>  <date>                        <dbl>                           <dbl>
#>  1 NH0910 1975-01-01                     21                              31  
#>  2 NH0910 1975-01-02                     21.2                            28.5
#>  3 NH0910 1975-01-03                     24.2                            32.5
#>  4 NH0910 1975-01-04                     23.5                            31  
#>  5 NH0910 1975-01-05                     19                              22  
#>  6 NH0910 1975-01-06                     22.2                            27.5
#>  7 NH0910 1975-01-07                     24.2                            30  
#>  8 NH0910 1975-01-08                     26.5                            33.5
#>  9 NH0910 1975-01-09                     26.5                            35  
#> 10 NH0910 1975-01-10                     19.2                            25  
#> # ℹ 15,543 more rows
#> # ℹ 31 more variables: temperatura_abrigo_150cm_minima <dbl>,
#> #   temperatura_intemperie_5cm_minima <lgl>,
#> #   temperatura_intemperie_50cm_minima <lgl>,
#> #   temperatura_suelo_5cm_media <lgl>, temperatura_suelo_10cm_media <lgl>,
#> #   temperatura_inte_5cm <lgl>, temperatura_intemperie_150cm_minima <lgl>,
#> #   humedad_suelo <lgl>, precipitacion_pluviometrica <dbl>, granizo <lgl>, …