Resumen Estadístico de Temperaturas por Estación

Calcula un resumen estadístico clave para la columna de temperatura (temperatura_abrigo_150cm) en el conjunto de datos de una o varias estaciones.

Usage

tabla_resumen_temperatura(estacion)

Arguments

estacion

El conjunto de datos de estaciones meteorológicas o de una estacion en especifico. Se utilizaran las columnas:

id: Identificador de la estacion.

temperatura_abrigo_150cm: Columna con los valores de temperatura utilizados para los cálculos.

Value

Un objeto tibble que contiene una fila por cada estación (id) con las siguientes columnas de resumen: id, Max, Min, Promedio, y Desvio_Estandar.

Details

La tabla de salida incluye la temperatura máxima, mínima, promedio y la desviación estándar. Los cálculos se agrupan por el identificador de la estación (id).

Esta función requiere el paquete dplyr para la manipulación de datos.

Examples

#> El archivo ya existe en la ruta indicada. Leyendo archivo...
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)
#> Rows: 18929 Columns: 35
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (2): id, direccion_viento_1000cm
#> dbl  (22): temperatura_abrigo_150cm, temperatura_abrigo_150cm_maxima, temper...
#> lgl  (10): temperatura_suelo_10cm_media, temperatura_inte_5cm, temperatura_i...
#> date  (1): fecha
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Lectura completada. El dataset tiene 18929 filas y 35 columnas.
#> El archivo ya existe en la ruta indicada. Leyendo archivo...
#> Rows: 15553 Columns: 35
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (1): id
#> dbl   (5): temperatura_abrigo_150cm, temperatura_abrigo_150cm_maxima, temper...
#> lgl  (28): temperatura_intemperie_5cm_minima, temperatura_intemperie_50cm_mi...
#> date  (1): fecha
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Lectura completada. El dataset tiene 15553 filas y 35 columnas.

tabla_resumen_temperatura(NH0437)
#> # A tibble: 1 × 5
#>   id       Max   Min Promedio Desvio_Estandar
#>   <chr>  <dbl> <dbl>    <dbl>           <dbl>
#> 1 NH0437  36.3   2.1     20.2            5.77
tabla_resumen_temperatura(NH0910)
#> # A tibble: 1 × 5
#>   id       Max   Min Promedio Desvio_Estandar
#>   <chr>  <dbl> <dbl>    <dbl>           <dbl>
#> 1 NH0910  35.2     0     17.7            6.12